requirements.txt
alembic==1.16.5
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pokebase==1.4.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
